chore: Memo & Rename
